trading_platform/requirements.txt
Django==5.0.1
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
Pillow==10.2.0
openai==1.3.7
django-taggit==5.0.1
django-allauth==0.60.1
stripe==7.11.0
python-slugify==8.0.1
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9